refactor(getMetadataFromStatement): Turned one file into module
Turned one file `getMetadataFromStatement` into module with a number of
existing metadata parse options and ability to add future ones

Closes LL-260
