+++ src/apps/statements/service/storeStatements/queriables/getMetadataFromStatement/index.ts
@@ -1,0 +1,19 @@
+import Statement from '../../../../models/Statement';
+import { getDurationMetadata } from './getDurationMetadata';
+import { getSequencingMetadata } from './getSequencingMetadata';
+
+export default (statement: Statement): {readonly [key: string]: any} => {
+  const durationMetadata = getDurationMetadata(statement);
+
+  if (durationMetadata) {
+    return durationMetadata;
+  }
+
+  const sequencingMetadata = getSequencingMetadata(statement);
+
+  if (sequencingMetadata) {
+    return sequencingMetadata;
+  }
+
+  return {};
+};
